pyproject.toml
[project]
name = "hyperliquid-whalealert-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.8.0",
    "python-dotenv>=1.1.0",
]
